Moving @ray.remote functions from graph.py and local_edges.py to utils.py file. Also, renaming test.py to genomics_example.py for clarity
